requirements.txt
Django
gunicorn
psycopg2-binary
mixer
pytest
pytest-django
pytest-cov
djangorestframework
django-rest-swagger